+++ library/Mockery/CompositeExpectation.php
@@ -43,9 +43,20 @@ class CompositeExpectation implements ExpectationInterface
     /**
      * @param mixed ...
      */
-    public function andReturn()
+    public function andReturn(...$args)
     {
-        return $this->__call(__FUNCTION__, func_get_args());
+        return $this->__call(__FUNCTION__, $args);
+    }
+
+    /**
+     * Set a return value, or sequential queue of return values
+     *
+     * @param mixed ...
+     * @return self
+     */
+    public function andReturns(...$args)
+    {
+        return call_user_func_array([$this, 'andReturn'], $args);
     }
 
     /**
@@ -104,9 +115,8 @@ class CompositeExpectation implements ExpectationInterface
      * @param mixed ...
      * @return \Mockery\Expectation
      */
-    public function shouldReceive()
+    public function shouldReceive(...$args)
     {
-        $args = func_get_args();
         reset($this->_expectations);
         $first = current($this->_expectations);
         return call_user_func_array(array($first->getMock(), 'shouldReceive'), $args);