+++ library/Mockery/Container.php
@@ -57,7 +57,7 @@ class Container
     protected $_groups = array();
 
     /**
-     * @var Generator\Generator
+     * @var Generator
      */
     protected $_generator;
 
@@ -85,17 +85,17 @@ class Container
      * names or partials - just so long as it's something that can be mocked.
      * I'll refactor it one day so it's easier to follow.
      *
+     * @param array $args
+     *
+     * @return Mock
      * @throws Exception\RuntimeException
-     * @throws Exception
-     * @return \Mockery\Mock
      */
-    public function mock()
+    public function mock(...$args)
     {
         $expectationClosure = null;
         $quickdefs = array();
         $constructorArgs = null;
         $blocks = array();
-        $args = func_get_args();
 
         if (count($args) > 1) {
             $finalArg = end($args);
@@ -122,16 +122,6 @@ class Container
             // check for multiple interfaces
             if (is_string($arg) && strpos($arg, ',') && !strpos($arg, ']')) {
                 $interfaces = explode(',', str_replace(' ', '', $arg));
-                foreach ($interfaces as $i) {
-                    if (!interface_exists($i, true) && !class_exists($i, true)) {
-                        throw new \Mockery\Exception(
-                            'Class name follows the format for defining multiple'
-                            . ' interfaces, however one or more of the interfaces'
-                            . ' do not exist or are not included, or the base class'
-                            . ' (which you may omit from the mock definition) does not exist'
-                        );
-                    }
-                }
                 $builder->addTargets($interfaces);
                 array_shift($args);
 
@@ -162,7 +152,7 @@ class Container
                 $builder->setWhiteListedMethods($partialMethods);
                 array_shift($args);
                 continue;
-            } elseif (is_string($arg) && (class_exists($arg, true) || interface_exists($arg, true))) {
+            } elseif (is_string($arg) && (class_exists($arg, true) || interface_exists($arg, true) || trait_exists($arg, true))) {
                 $class = array_shift($args);
                 $builder->addTarget($class);
                 continue;
@@ -421,7 +411,7 @@ class Container
      * Store a mock and set its container reference
      *
      * @param \Mockery\Mock
-     * @return \Mockery\Mock
+     * @return \Mockery\MockInterface
      */
     public function rememberMock(\Mockery\MockInterface $mock)
     {