+++ library/Mockery/Generator/StringManipulationGenerator.php
@@ -21,10 +21,45 @@
 namespace Mockery\Generator;
 
 use Mockery\Generator\StringManipulation\Pass\Pass;
+use Mockery\Generator\StringManipulation\Pass\RemoveDestructorPass;
+use Mockery\Generator\StringManipulation\Pass\CallTypeHintPass;
+use Mockery\Generator\StringManipulation\Pass\MagicMethodTypeHintsPass;
+use Mockery\Generator\StringManipulation\Pass\ClassNamePass;
+use Mockery\Generator\StringManipulation\Pass\ClassPass;
+use Mockery\Generator\StringManipulation\Pass\TraitPass;
+use Mockery\Generator\StringManipulation\Pass\InstanceMockPass;
+use Mockery\Generator\StringManipulation\Pass\InterfacePass;
+use Mockery\Generator\StringManipulation\Pass\MethodDefinitionPass;
+use Mockery\Generator\StringManipulation\Pass\RemoveBuiltinMethodsThatAreFinalPass;
+use Mockery\Generator\StringManipulation\Pass\RemoveUnserializeForInternalSerializableClassesPass;
+use Mockery\Generator\StringManipulation\Pass\AvoidMethodClashPass;
 
 class StringManipulationGenerator implements Generator
 {
     protected $passes = array();
+
+    /**
+     * Creates a new StringManipulationGenerator with the default passes
+     *
+     * @return StringManipulationGenerator
+     */
+    public static function withDefaultPasses()
+    {
+        return new static([
+            new CallTypeHintPass(),
+            new MagicMethodTypeHintsPass(),
+            new ClassPass(),
+            new TraitPass(),
+            new ClassNamePass(),
+            new InstanceMockPass(),
+            new InterfacePass(),
+            new AvoidMethodClashPass(),
+            new MethodDefinitionPass(),
+            new RemoveUnserializeForInternalSerializableClassesPass(),
+            new RemoveBuiltinMethodsThatAreFinalPass(),
+            new RemoveDestructorPass(),
+        ]);
+    }
 
     public function __construct(array $passes)
     {