+++ library/Mockery/Matcher/MultiArgumentClosure.php
@@ -20,7 +20,7 @@
 
 namespace Mockery\Matcher;
 
-class MultiArgumentClosure extends MatcherAbstract
+class MultiArgumentClosure extends MatcherAbstract implements ArgumentListMatcher
 {
 
     /**